requirements.txt
fastapi==0.114.1
pydantic==2.9.1
uvicorn==0.30.6
nltk==3.9.1
python-multipart==0.0.12
aiofiles==24.1.0
beautifulsoup4==4.12
langdetect==1.0.9
numpy==1.26.4
torch --index-url https://download.pytorch.org/whl/cu118
transformers
scikit-learn==1.5.2
scikit-learn-extra==0.3.0
yake==0.4.8
tk
pillow
Jinja2
playwright
